+++ app/src/main/java/com/example/intuitionproject/screens/homedashboard/chats/ChatFragment.java
@@ -1,18 +1,15 @@
-package com.example.intuitionproject.ui.chats;
+package com.example.intuitionproject.screens.homedashboard.chats;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.intuitionproject.R;
 import com.example.intuitionproject.databinding.FragmentChatBinding;
 
 public class ChatFragment extends Fragment {
