+++ app/src/main/java/com/example/intuitionproject/screens/homedashboard/chats/ChatViewModel.java
@@ -1,16 +1,16 @@
-package com.example.intuitionproject.ui.profile;
+package com.example.intuitionproject.screens.homedashboard.chats;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ProfileViewModel extends ViewModel {
+public class ChatViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public ProfileViewModel() {
+    public ChatViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is dashboard fragment");
     }
 
     public LiveData<String> getText() {
