+++ app/src/main/java/com/example/intuitionproject/screens/homedashboard/profile/ProfileViewModel.java
@@ -1,16 +1,16 @@
-package com.example.intuitionproject.ui.chats;
+package com.example.intuitionproject.screens.homedashboard.profile;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ChatViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public ChatViewModel() {
+    public ProfileViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is notifications fragment");
     }
 
     public LiveData<String> getText() {
