+++ app/src/main/res/layout/fragment_chat.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.chats.ChatFragment">
+    tools:context=".screens.homedashboard.chats.ChatFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler"
