+++ app/src/main/res/layout/view_pager_listing_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.ViewPagerListing">
+    tools:context=".screens.homedashboard.home.ViewPagerListing">
 
     <ImageView
         android:id="@+id/imageView"
